+++ src/main/java/config/WorldConfig.java
@@ -10,7 +10,7 @@ public class WorldConfig {
     public int meso_rate = 1;
     public int drop_rate = 1;
     public int boss_drop_rate = 1;
-    public int quest_rate = 1;
+    public double quest_rate = 1;
     public int travel_rate = 1;
     public int fishing_rate = 1;
 }
